+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/AdmissionRegistrationAPIGroupDSL.java
@@ -19,6 +19,4 @@ public interface AdmissionRegistrationAPIGroupDSL extends Client {
   V1AdmissionRegistrationAPIGroupDSL v1();
 
   V1beta1AdmissionRegistrationAPIGroupDSL v1beta1();
-
-  V1Alpha1AdmissionRegistrationAPIGroupDSL v1alpha1();
 }